+++ src/HoverProvider.ts
@@ -20,9 +20,6 @@ export class HoverProvider implements vscode.HoverProvider {
      */
     public provideHover(document: vscode.TextDocument, position: vscode.Position, token: vscode.CancellationToken): Thenable<vscode.Hover> {
         const settings = vscode.workspace.getConfiguration('asm-code-lens');
-        if(settings.enableHovering == false) 
-            return undefined;
-
         return this.search(document, position);
     }