+++ src/extension.ts
@@ -8,51 +8,163 @@ import { CodeLensProvider } from './CodeLensProvider';
 import { RenameProvider } from './RenameProvider';
 import { CompletionProposalsProvider } from './CompletionProposalsProvider';
 import { Commands } from './Commands';
+import { setGrepGlobPatterns } from './grep';
 
 export function activate(context: vscode.ExtensionContext) {
-    const settings = vscode.workspace.getConfiguration('asm-code-lens');
+	// Enable logging.
+    configure(context);
     
-    // Note: don't add 'language' property, otherwise other extension with similar file pattern may not work.
-    // If the identifier is missing it also don't help to define it in package.json. And if "id" would be used it clashes again with other extensions.
-    const asmFiles: vscode.DocumentSelector = { scheme: "file", pattern:"**/*.{asm,s,a80,inc}"};
-
-    if(settings.enableGotoDefinition != false) {
-        context.subscriptions.push(
-            vscode.languages.registerCodeLensProvider(asmFiles, new CodeLensProvider()),
-        );
-    }
-
-    context.subscriptions.push(
-        vscode.languages.registerHoverProvider(asmFiles, new HoverProvider()),
-    );
-
-    context.subscriptions.push(
-        vscode.languages.registerCompletionItemProvider(asmFiles, new CompletionProposalsProvider()),
-    );
-
-    if(settings.enableGotoDefinition != false) {
-            context.subscriptions.push(
-            vscode.languages.registerDefinitionProvider(asmFiles, new DefinitionProvider())
-        );
-    }
+    // Check for every change.
+	context.subscriptions.push(vscode.workspace.onDidChangeConfiguration(event => {
+        configure(context, event);
+    }));
     
-    context.subscriptions.push(
-        //vscode.languages.registerReferenceProvider(ASM_LANGUAGE, new ReferenceProvider())
-        vscode.languages.registerReferenceProvider(asmFiles, new ReferenceProvider())
-    );
-
-    if(settings.enableRenaming != false) {
-        context.subscriptions.push(
-            vscode.languages.registerRenameProvider(asmFiles, new RenameProvider()),
-        );
-    }
-
-
+    // Register command once.
     vscode.commands.registerCommand('asm-code-lens.find-labels-with-no-reference', () => {
         Commands.findLabelsWithNoReference(); 
     });
-   
 }
+ 
+
+/**
+ * Reads the confguration.
+ */
+function configure(context: vscode.ExtensionContext, event?) {
+    const settings = vscode.workspace.getConfiguration('asm-code-lens');
+    
+    // Check if grep paths have changed
+    if(event) {
+        if(event.affectsConfiguration('asm-code-lens.includeFiles')
+            || event.affectsConfiguration('asm-code-lens.excludeFiles')) {
+            // Restart provider if running:
+            if(regCodeLensProvider) {
+                // Deregister
+                regCodeLensProvider.dispose();
+                regCodeLensProvider = undefined;
+            }            
+            if(regReferenceProvider) {
+                // Deregister
+                regReferenceProvider.dispose();
+                regReferenceProvider = undefined;
+            }
+            if(regRenameProvider) {
+                // Deregister
+                regRenameProvider.dispose();
+                regRenameProvider = undefined;
+            }
+        }
+    }
+
+    // Set search paths.
+    setGrepGlobPatterns(settings.includeFiles, settings.excludeFiles);
+
+
+    // Note: don't add 'language' property, otherwise other extension with similar file pattern may not work.
+    // If the identifier is missing it also don't help to define it in package.json. And if "id" would be used it clashes again with other extensions.
+    const asmFiles: vscode.DocumentSelector = {scheme: "file", pattern: settings.includeFiles};
+
+    // Code Lenses
+    if(settings.enableCodeLenses) {
+        if(!regCodeLensProvider) {
+            // Register
+            regCodeLensProvider = vscode.languages.registerCodeLensProvider(asmFiles, new CodeLensProvider());
+            context.subscriptions.push(regCodeLensProvider);
+        }
+    }
+    else {
+        if(regCodeLensProvider) {
+            // Deregister
+            regCodeLensProvider.dispose();
+            regCodeLensProvider = undefined;
+        }
+    }
+
+    if(settings.enableHovering) {
+        if(!regHoverProvider) {
+            // Register
+            regHoverProvider = vscode.languages.registerHoverProvider(asmFiles, new HoverProvider());
+            context.subscriptions.push(regHoverProvider);
+        }
+    }
+    else {
+        if(regHoverProvider) {
+            // Deregister
+            regHoverProvider.dispose();
+            regHoverProvider = undefined;
+        }
+    }
+
+    if(settings.enableCompletions) {
+        if(!regCompletionProposalsProvider) {
+            // Register
+            regCompletionProposalsProvider = vscode.languages.registerCompletionItemProvider(asmFiles, new CompletionProposalsProvider());
+            context.subscriptions.push(regCompletionProposalsProvider);
+        
+        }
+    }
+    else {
+        if(regCompletionProposalsProvider) {
+            // Deregister
+            regCompletionProposalsProvider.dispose();
+            regCompletionProposalsProvider = undefined;
+        }
+    }
+
+    if(settings.enableGotoDefinition) {
+        if(!regDefinitionProvider) {
+            // Register
+            regDefinitionProvider = vscode.languages.registerDefinitionProvider(asmFiles, new DefinitionProvider());
+            context.subscriptions.push(regDefinitionProvider);
+        }
+    }
+    else {
+        if(regDefinitionProvider) {
+            // Deregister
+            regDefinitionProvider.dispose();
+            regDefinitionProvider = undefined;
+        }
+    }
+    
+    if(settings.enableFindAllReferences) {
+        if(!regReferenceProvider) {
+            // Register
+            //vscode.languages.registerReferenceProvider(ASM_LANGUAGE, new ReferenceProvider())   
+            regReferenceProvider = vscode.languages.registerReferenceProvider(asmFiles, new ReferenceProvider());
+            context.subscriptions.push(regReferenceProvider);
+        }
+    }
+    else {
+        if(regReferenceProvider) {
+            // Deregister
+            regReferenceProvider.dispose();
+            regReferenceProvider = undefined;
+        }
+    }
+
+    if(settings.enableRenaming) {
+        if(!regRenameProvider) {
+            // Register
+            regRenameProvider = vscode.languages.registerRenameProvider(asmFiles, new RenameProvider());
+            context.subscriptions.push(regRenameProvider);
+        }
+    }
+    else {
+        if(regRenameProvider) {
+            // Deregister
+            regRenameProvider.dispose();
+            regRenameProvider = undefined;
+        }
+    }
+}
+let regCodeLensProvider;
+let regHoverProvider;
+let regCompletionProposalsProvider;
+let regDefinitionProvider;
+let regReferenceProvider;
+let regRenameProvider;
+
+
+
 
 // this method is called when your extension is deactivated
 export function deactivate() {