+++ CHANGELOG.md
@@ -1,8 +1,10 @@
 # Changelog
 
 ## 1.3.0
-- Settings
+- Settings fixed.
+- Added settings for globbing include and exclude files.
 - Fixed error reading files, sometimes wrong data was read, i.e. sometimes labels were not found.
+
 
 ## 1.2.2
 - Added Z80 register names as proposals for completions.